Fairings with Lips
As we discussed earlier, I went ahead to try my hand at making the lips
on the fairings to get an overlap and ensure a smooth continuous
surface. I added .5mm to the overall thickness of the shell to aid in
stability while printing tall parts and to help make the lip possible.